ViewController now contains all controller methods
